Make the way that the ef Parser received notes more generic. (#12)
* Add MusicSource and StringMusicSource

* Parser takes a MusicSource instead of String[]

* Fix junit parameterised tests running in gradle

* Change music index to long instead of int to support longer pieces

* Add MusicSource which reads from files without pulling the whole piece into memory.  Now we can play GIANT pieces with constant memory.

* Make ef-core Main play a file by default

+++ vibe/src/main/java/eflang/vibe/CodePane.java
@@ -1,10 +1,9 @@
 package eflang.vibe;
 
-import java.awt.Color;
-
-import javax.swing.JTextPane;
+import javax.swing.*;
 import javax.swing.text.Style;
 import javax.swing.text.StyleConstants;
+import java.awt.*;
 
 public class CodePane extends JTextPane {
     private char mDivider;
@@ -22,7 +21,7 @@ public class CodePane extends JTextPane {
         return mDivider;
     }
 
-    public void setCurrentCommandIndex(int i) {
+    public void setCurrentCommandIndex(long i) {
         Style style = this.addStyle("Redtext", null);
         StyleConstants.setForeground(style, Color.red);
         Style noStyle = this.addStyle("NoStyle", null);
